add test for response with newlines instead of \r\n
While the HTTP standard specifies that the status-line and header-fields
should be terminated with a CRLF (\r\n), many implementations on the
internet would wrongly output only LF (\n). This is even recognized in the
RFC which  suggests that a LF may be recognized as a line terminator.

https://tools.ietf.org/html/rfc7230#section-3.5
>   Although the line terminator for the start-line and header fields is
>   the sequence CRLF, a recipient MAY recognize a single LF as a line
>   terminator and ignore any preceding CR.

This commit introduces a test for a response with only LF newlines.