Don't add the Host header if it already exists
the request can already contain a host header, either because it came
from the async_h1::accept, or because the user set it manually. In that
case don't add a second host header (the one based on `req.url()`).

Update http-types to "1.2.0", because it's the first version that
introduced `headers::HOST`.