Replace async-std with smol crates
It would be nice to be able to use this in environments where I use
smol. However this crate imports the async-std crate for I/O. This
commit partially replaces async-std with the equivalent combinators from
futures-lite and async-io, two smol crates that async-std depends on
anyways.

The task spawn is still left in, as replacing it will be controversial.

Signed-off-by: John Nunley <dev@notgull.net>